wtp_3. rel_err h3 fixed. Writed some tp sections.
