+++ doc/latex_ben_jose_tool_paper/.ispell_english
@@ -1,45 +1,73 @@
+aa
+api
+argc
+args
+argv
+assig
+basicstyle
 BCFF
 BCFFs
 BCP
 BDUST
 Beltran
 ben
+bj
+bjr
+breaklines
 CDCL
 cnf
+const
 da
-diff
 DIMACS
 DPLL
 EasyChair
 EPiC
 ER
 Federman
+ff
 fibres
 github
 glb
+html
+http
 https
+int
 Joao
 jose
 joseluisquiroga
 leveldat
+libben
 Magda
 Marquez
 neuromap
 NP
-pdflscape
-positon
+numberstyle
+oo
+org
+printf
+PTH
 quanton
-quantons
 Quiroga
+rdynamic
 Rios
+satcompetition
+satisf
 sha
 sorset
 sortee
 sortees
 sortor
 sortrel
+ss
+std
+tabsize
 tauto
 Tseitin
+UNS
+val
+vv
+www
+xleftmargin
 yahoo
 Yashua
 YHWH
